experiment: snapshot vs reference for board entities
Currently trying to figure out how to organize the
Board.java class. I tried approaching it with snapshots
of the entities but I seem to have encountered an issue.

Currently this branch is just for saving any attempts at
making the board structure better, and to see which would
work best for this project. I expect that the reference
will likely be my approach, but I'll have to see.